refactor: Runtimes
All roles can see own clusters of runtime
Remove dead codes in runtime store
Refine root store
Remove useless page components
