--- src/config/runtime-providers.js
@@ -1,7 +1,0 @@
-export default {
-  // key => label
-  qingcloud: 'QingCloud',
-  aws: 'AWS',
-  aliyun: 'Aliyun',
-  kubernetes: 'Kubernetes'
-};
