--- src/pages/Dashboard/Runtimes/Add/index.scss
@@ -1,135 +1,0 @@
-@import '~scss/vars';
-
-.createForm {
-  display: block;
-  //padding: 36px 0 0 36px;
-
-  &>div{
-    margin-bottom: 24px;
-    label{
-      margin-bottom: 0;
-    }
-  }
-  &>.showDiv{
-    margin-bottom: 22px;
-    .rightShow{
-      margin-top: 7px;
-    }
-    .showValue{
-      color: $text-color;
-    }
-  }
-  .zoneItem{
-    height: 32px;
-    >label{
-      float: left;
-      line-height: 32px;
-    }
-  }
-  .name{
-    display: inline-block;
-    //float: left;
-    margin-right: 24px;
-    width: 168px;
-    line-height: 16px;
-    font-size: 14px;
-    font-weight: 500;
-    color: $N500;
-    text-align: right;
-  }
-  .fl{
-    float: left;
-    line-height: 32px;
-  }
-  .input{
-    width: 256px;
-  }
-  .inputUrl{
-    width: 376px;
-  }
-  .inputSmall{
-    width: 138px;
-    margin-right: 10px;
-  }
-  .inputTitle{
-    display: inline-block;
-    margin: 8px 12px 3px 0;
-    width: 124px;
-    font-size: 12px;
-    color: $N100;
-  }
-  .inputMiddle{
-    width: 124px;
-    margin-right: 12px;
-  }
-  .radio,.checkbox{
-    margin-right: 30px;
-  }
-  .add{
-    display: inline-block;
-  }
-  .addBottom{
-    margin-left: 192px;
-  }
-  .select{
-    width: 256px;
-    margin-right: 12px;
-  }
-  .textareaItem{
-    height: 88px;
-  }
-  .textarea{
-    box-sizing: border-box;
-    padding: 7px 10px;
-    width: 400px;
-    max-width: calc(100% - 192px);
-    height: 88px;
-    border-radius: 2px;
-    background-color: $N0;
-    border: 1px solid $N30;
-    outline: none;
-    //resize: none;
-    &:focus{
-      box-shadow: 0 1px 2px 0 rgba(109, 53, 195, 0.15);
-      border: solid 1px $P75;
-    }
-  }
-  .textareaName{
-    position: relative;
-    bottom: 70px;
-  }
-
-  .rightShow{
-    margin-top: 8px;
-    margin-left: 192px;
-    font-size: 12px;
-    line-height: 1.67;
-    color: $N100;
-  }
-
-  .submitBtnGroup{
-    margin: 0 -32px;
-    padding: 16px 0 16px 224px;
-    border-top: 1px solid $N10;
-    margin-bottom: 0;
-    button{
-      margin-right: 12px;
-    }
-  }
-  .credentialTip {
-    margin-top: 10px;
-    font-size:12px;
-    color: $N100;
-    margin-left: 192px;
-  }
-
-  .radioGroup {
-    label {
-      float: left;
-    }
-  }
-}
-
-
-
-
