--- src/pages/Dashboard/Runtimes/Detail/index.scss
@@ -1,1 +1,0 @@
-@import '~scss/vars';
