+++ src/pages/Dashboard/Runtimes/index.scss
@@ -1,2 +1,294 @@
-@import '~scss/vars';
+@import "~scss/vars";
+
+$body-margin: 132px;
+$pg-margin: 18px;
+$card-height: 188px;
+
+%btnAdd {
+  margin-top: 20px;
+  :global {
+    .icon {
+      vertical-align: bottom;
+    }
+  }
+}
+
+.layout {
+  margin-top: $pg-margin !important;
+
+  // fix normal layout
+  &.isNormal {
+    margin-top: 0 !important;
+    .leftPanel {
+      max-width: 220px;
+    }
+    .page {
+      margin: 0 auto;
+    }
+  }
+
+  .leftPanel {
+    max-width: 205px;
+    border-right: 1px solid $N10;
+  }
+
+  // fix left/right panel width not match design
+  @media screen and (min-width: 1200px) {
+    .rightPanel {
+      width: 885px;
+    }
+  }
+
+  .pageTitle {
+    font-size: $size-medium;
+    color:$N500;
+    margin-bottom: $pg-margin;
+    margin-left: 102px;
+  }
+}
+
+.page {
+  display: block;
+  margin-left: $body-margin - 30px;
+  margin-right: $body-margin - 30px;
+
+  .platforms {
+    border-right: 1px solid $N10;
+    .provider {
+      height: 40px;
+      padding-left: 12px;
+      cursor: pointer;
+
+      &.active, &:hover {
+        box-shadow: 0 1px 0 0 $N10, 0 -1px 0 0 $N10;
+        background-color: $N0;
+      }
+      &.disabled {
+        opacity: 0.5;
+        cursor: default;
+        box-shadow: none;
+        background-color: transparent;
+      }
+
+      span {
+        display: inline-block;
+        padding: 12px 0 !important;
+        vertical-align: middle;
+
+        &.proName {
+          font-size: $size-small;
+          margin-left: 8px;
+        }
+        &.proCount {
+          font-size: $size-small;
+          float: right;
+          margin-right: 23px;
+        }
+      }
+    }
+
+    .tooltip {
+      width: 100%;
+      cursor: default;
+    }
+    .popper {
+      margin-bottom: 0;
+    }
+  }
+
+  .tabs {
+    text-align: center;
+    background: none;
+    border: none;
+    margin-top: -40px;
+    margin-bottom: 12px;
+
+    &>div {
+      font-size: $size-normal;
+      line-height: 1.14;
+      padding-bottom: 10px;
+      margin-right: 20px;
+    }
+  }
+
+  .body {
+    padding: 0;
+
+    .emptyData {
+      height: 303px;
+      @include centerAll;
+
+      p:first-child {
+        font-size: $size-normal;
+        font-weight: 500;
+        line-height: 2;
+        color: $N500;
+
+        & + p {
+          font-size: $size-small;
+          line-height: 1.67;
+          color: $N75;
+        }
+      }
+      .btnAddEnv {
+        @extend %btnAdd;
+      }
+    }
+
+    .actionPop {
+      .actionBar {
+        transform: rotate(90deg);
+      }
+      :global {
+        .operate-menu {
+          font-size: $size-small;
+          .icon {
+            display: inline-block;
+            min-width: auto;
+            background-color: transparent;
+            padding-left: 0;
+            padding-right: 4px;
+            vertical-align: sub;
+          }
+        }
+      }
+    }
+
+    .fixMenu {
+      margin-top: -10px;
+    }
+
+    .envs {
+      flex-flow: wrap;
+
+      .envItem {
+        height: $card-height;
+        .name {
+          display: inline-block;
+          font-size: $size-normal;
+          font-weight: 500;
+          line-height: 2;
+          color: $N500;
+        }
+        .desc {
+          display: block;
+          font-size: $size-small;
+          line-height: 1.67;
+          color: $N75;
+        }
+
+        .bottomInfo {
+          margin-top: 40px;
+          .info {
+            display: inline-block;
+            margin-right: 20px;
+
+            .label {
+              font-size: $size-small;
+              line-height: 1.67;
+              color: $N75;
+            }
+            .val {
+              font-size: $size-small;
+              font-weight: 500;
+              line-height: 1.67;
+              color: $N500;
+            }
+          }
+        }
+      }
+
+      .cardAddEnv {
+        height: $card-height;
+        background: $N10;
+        @include centerAll;
+
+        .btnAdd {
+          @extend %btnAdd;
+          color: $N300;
+        }
+      }
+
+    }
+
+    .authInfos {
+      .info {
+        padding: 14px auto;
+        font-size: $size-small;
+        &>span {
+          display: inline-block;
+          margin-right: 20px;
+        }
+        .name {
+          font-weight: 500;
+          color: $N300;
+          min-width: 50px;
+          width: 20%;
+        }
+        .desc{
+          min-width: 100px;
+          width: 20%;
+          color: $N75;
+        }
+        .cont {
+          min-width: 140px;
+          color: $N75;
+        }
+        .time {
+          float: right;
+          margin-right: 64px;
+          color: $N65;
+        }
+      }
+      .addAuthInfo {
+        margin-top: 25px;
+        :global {
+          .icon {
+            vertical-align: sub;
+          }
+        }
+        &>.btnAdd {
+          font-size: $size-small;
+          color: $P75;
+          cursor: pointer;
+          margin-left: 4px;
+        }
+      }
+    }
+  }
+}
+
+
+// modify runtime and credential
+.fmCtrl {
+  margin-bottom: 24px;
+  .label {
+    display: block;
+    font-weight: 500;
+    color: $N500;
+    margin-bottom: 8px;
+  }
+  .field {
+    height: 32px;
+    border-radius: 2px;
+    background-color: $N0;
+  }
+}
+
+.dialog {
+  .item {
+    padding: 10px;
+    border-radius: 2px;
+    box-shadow: 0 1px 4px 0 rgba(73, 33, 173, 0.06), 0 4px 8px 0 rgba(35, 35, 36, 0.04);
+    background-color: $N0;
+    cursor: pointer;
+    &:hover, &.checked {
+      border: 1px solid $P75;
+    }
+    .icon {
+      float: right;
+    }
+  }
+}
+
